Fix Float32 -> wide Decimal (Decimal128/256) via Float64 to avoid wide-integer narrowing
Casting a `Float32` whose scaled value lands above the `Int64` range but within
the target's native range produced garbage: `static_cast<ToNativeType>(out)` on a
`Float32` picks `wide::integer`'s templated builtin-float constructor, which
narrows through `int64_t` in `wide_integer_from_builtin`. This is a pre-existing
bug (present in stock, since master does the same `static_cast<ToNativeType>(out)`),
surfaced while reworking this conversion.

```
select CAST(toFloat32(1), 'Decimal128(20)'), CAST(toFloat32(1), 'Decimal256(30)');

was
┌─CAST(toFloat⋯al128(20)')─┬─CAST(toFloat32(1), 'Decimal256(30)')─┐
│   0.09223372036854775807 │     0.000000000009223372036854775807 │
└──────────────────────────┴──────────────────────────────────────┘

now
┌─CAST(toFloat⋯al128(20)')─┬─CAST(toFloat32(1), 'Decimal256(30)')─┐
│   1.00000002004087734275 │     1.000000015047466219891198222336 │
└──────────────────────────┴──────────────────────────────────────┘
```

Widen the float result to `Float64` before the integer cast in both the scalar
(`convertToDecimalImpl`) and batch (`convertToDecimalBatch`) paths. `Float64` ->
wide integer is safe; the widen is a no-op for `Float64` sources and exact for
`Float32`, so only the buggy `Float32` -> `Decimal128`/`Decimal256` path changes.

Add a regression to `04034_decimal_cast_float_rounding` (scalar and batch).

+++ src/DataTypes/DataTypesDecimal.cpp
@@ -454,7 +454,7 @@ ReturnType convertToDecimalImpl(const typename FromDataType::FieldType & value, 
                 return ReturnType(false);
         }
 
-        result = static_cast<ToNativeType>(out);
+        result = static_cast<ToNativeType>(static_cast<Float64>(out));
         return ReturnType(true);
     }
     else
@@ -529,10 +529,15 @@ NO_SANITIZE_UNDEFINED void convertToDecimalBatch(
                 if constexpr (!tight)
                     overflow = overflow || !isFinite(out);
 
+                /// Widen to `Float64` before the integer cast: `wide::integer`'s builtin-float
+                /// constructor narrows a `Float32` through `int64_t`, so a `Float32` value above the
+                /// `Int64` range (e.g. `Float32` -> `Decimal128`/`Decimal256` at scale > 18) would be
+                /// cast incorrectly. `Float64` -> wide integer is safe; the widen is a no-op for
+                /// `Float64` and exact for `Float32`.
                 if constexpr (has_nullmap)
                 {
                     nullmap_ptr[i] = overflow;
-                    to_ptr[i] = overflow ? static_cast<ToNativeType>(0) : static_cast<ToNativeType>(out);
+                    to_ptr[i] = overflow ? static_cast<ToNativeType>(0) : static_cast<ToNativeType>(static_cast<Float64>(out));
                 }
                 else
                 {
@@ -547,7 +552,7 @@ NO_SANITIZE_UNDEFINED void convertToDecimalBatch(
                                 "{} convert overflow. Float is out of Decimal range",
                                 ToDataType::family_name);
                     }
-                    to_ptr[i] = static_cast<ToNativeType>(out);
+                    to_ptr[i] = static_cast<ToNativeType>(static_cast<Float64>(out));
                 }
             }
         };
